优
更改数据表价格精度，注册正则校验，购物车选择性结算清除，搜索无商品时样式

+++ Farm/src/main/java/com/ljh/farm/entity/ProductDetail.java
@@ -27,10 +27,10 @@ public class ProductDetail extends Base{
     private String img;
 
     @TableField("pri")
-    private Integer pri;
+    private Double pri;
 
     @TableField("old_pri")
-    private Integer oldPri;
+    private Double oldPri;
 
     @TableField("discount")
     private String discount;
